% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{bs_button_group}
\alias{bs_button_group}
\alias{bs_button_toolbar}
\title{Bootstrap button group / toolbar}
\usage{
bs_button_group(..., size = NULL, vertical = FALSE, label = NULL, class = NULL)

bs_button_toolbar(..., label = NULL, class = NULL)
}
\arguments{
\item{...}{Buttons (\code{\link[=bs_button]{bs_button()}}) and named HTML attributes.}

\item{size}{\code{"sm"} or \code{"lg"}.}

\item{vertical}{If \code{TRUE}, stack vertically (\code{.btn-group-vertical}).}

\item{label}{Accessible label (\code{aria-label}).}

\item{class}{Extra classes.}
}
\value{
A button group / toolbar tag.
}
\description{
Bootstrap button group / toolbar
}
\examples{
bs_button_group(bs_button(label = "Left"), bs_button(label = "Right"))
bs_button_toolbar(
  bs_button_group(bs_button(label = "Cut")),
  bs_button_group(bs_button(label = "Copy"))
)
}
