% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{bs_container}
\alias{bs_container}
\title{Bootstrap container}
\usage{
bs_container(..., fluid = FALSE, breakpoint = NULL, class = NULL)
}
\arguments{
\item{...}{Content, and named HTML attributes.}

\item{fluid}{If \code{TRUE}, a full-width \code{.container-fluid}. Ignored when
\code{breakpoint} is set.}

\item{breakpoint}{One of \code{"sm"}, \code{"md"}, \code{"lg"}, \code{"xl"}, \code{"xxl"} for a
responsive \code{.container-{breakpoint}}.}

\item{class}{Extra classes.}
}
\value{
A container tag.
}
\description{
Bootstrap container
}
\examples{
bs_container(bs_row(bs_col("a"), bs_col("b")))
}
\seealso{
\code{\link[=bs_row]{bs_row()}}, \code{\link[=bs_col]{bs_col()}}
}
