% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{bs_display_heading}
\alias{bs_display_heading}
\title{Bootstrap display heading}
\usage{
bs_display_heading(..., level = 1, class = NULL)
}
\arguments{
\item{...}{Heading content and named HTML attributes.}

\item{level}{Display size / heading level (1-6).}

\item{class}{Extra classes.}
}
\value{
A heading tag.
}
\description{
A larger, slightly more opinionated heading style (\verb{.display-*}).
}
\examples{
bs_display_heading("Display heading", level = 2)
}
