% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{bs_lead}
\alias{bs_lead}
\title{Bootstrap lead paragraph}
\usage{
bs_lead(..., class = NULL)
}
\arguments{
\item{...}{Paragraph content and named HTML attributes.}

\item{class}{Extra classes.}
}
\value{
A paragraph tag.
}
\description{
A standout opening paragraph (\code{.lead}).
}
\examples{
bs_lead("This is a lead paragraph.")
}
