% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{bs_list_unstyled}
\alias{bs_list_unstyled}
\alias{bs_list_inline}
\title{Bootstrap unstyled / inline lists}
\usage{
bs_list_unstyled(..., class = NULL)

bs_list_inline(..., class = NULL)
}
\arguments{
\item{...}{List items (unnamed) and named HTML attributes for the \verb{<ul>}.}

\item{class}{Extra classes.}
}
\value{
An unordered list tag.
}
\description{
Remove a list's default styling (\code{.list-unstyled}) or lay items out inline
(\code{.list-inline}). Each unnamed \code{...} argument becomes one \verb{<li>}; named
arguments become attributes of the \verb{<ul>}.
}
\examples{
bs_list_unstyled("First", "Second", "Third")
bs_list_inline("One", "Two", "Three")
}
