% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviors.R
\name{bs_visually_hidden}
\alias{bs_visually_hidden}
\title{Visually hidden text}
\usage{
bs_visually_hidden(..., class = NULL)
}
\arguments{
\item{...}{Content and named HTML attributes.}

\item{class}{Extra classes.}
}
\value{
A \verb{<span>} tag.
}
\description{
Renders content available to assistive technologies but hidden from sighted
users (\code{.visually-hidden}).
}
\examples{
bs_visually_hidden("Loading, please wait")
}
