% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviors.R
\name{bs_vr}
\alias{bs_vr}
\title{Vertical rule}
\usage{
bs_vr(class = NULL)
}
\arguments{
\item{class}{Extra classes.}
}
\value{
A \verb{<div>} tag.
}
\description{
A vertical divider (\code{.vr}), the vertical counterpart of \verb{<hr>}.
}
\examples{
bs_vr()
}
