% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{show_bs_dropdown}
\alias{show_bs_dropdown}
\alias{hide_bs_dropdown}
\alias{toggle_bs_dropdown}
\title{Open, close or toggle a dropdown from the server}
\usage{
show_bs_dropdown(id, session = shiny::getDefaultReactiveDomain())

hide_bs_dropdown(id, session = shiny::getDefaultReactiveDomain())

toggle_bs_dropdown(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Dropdown id.}

\item{session}{The Shiny session.}
}
\value{
Nothing, called for their side effect.
}
\description{
Drives the \code{\link[=bs_dropdown]{bs_dropdown()}} or \code{\link[=bs_nav_dropdown]{bs_nav_dropdown()}} registered under \code{id}.
Its open state is reported back as \code{input$id}.
}
\examples{
if (interactive()) show_bs_dropdown("menu")
}
\seealso{
\code{\link[=bs_dropdown]{bs_dropdown()}}
}
