% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{show_bs_modal}
\alias{show_bs_modal}
\alias{hide_bs_modal}
\alias{toggle_bs_modal}
\title{Control a modal from the server}
\usage{
show_bs_modal(id, session = shiny::getDefaultReactiveDomain())

hide_bs_modal(id, session = shiny::getDefaultReactiveDomain())

toggle_bs_modal(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Modal id (namespaced automatically inside modules).}

\item{session}{The Shiny session.}
}
\value{
Invisibly \code{NULL}, called for its side effect.
}
\description{
Show, hide or toggle a \code{\link[=bs_modal]{bs_modal()}} from server code.
}
\examples{
if (interactive()) show_bs_modal("info")
if (interactive()) hide_bs_modal("info")
if (interactive()) toggle_bs_modal("info")
}
