% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offcanvas.R
\name{show_bs_offcanvas}
\alias{show_bs_offcanvas}
\alias{hide_bs_offcanvas}
\alias{toggle_bs_offcanvas}
\title{Control an offcanvas from the server}
\usage{
show_bs_offcanvas(id, session = shiny::getDefaultReactiveDomain())

hide_bs_offcanvas(id, session = shiny::getDefaultReactiveDomain())

toggle_bs_offcanvas(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Offcanvas id (namespaced automatically inside modules).}

\item{session}{The Shiny session.}
}
\value{
Invisibly \code{NULL}, called for its side effect.
}
\description{
Show, hide or toggle a \code{\link[=bs_offcanvas]{bs_offcanvas()}} from server code.
}
\examples{
if (interactive()) show_bs_offcanvas("menu")
if (interactive()) hide_bs_offcanvas("menu")
if (interactive()) toggle_bs_offcanvas("menu")
}
