% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toast.R
\name{show_bs_toast}
\alias{show_bs_toast}
\alias{hide_bs_toast}
\title{Control a toast from the server}
\usage{
show_bs_toast(id, session = shiny::getDefaultReactiveDomain())

hide_bs_toast(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Toast id (namespaced automatically inside modules).}

\item{session}{The Shiny session.}
}
\value{
Invisibly \code{NULL}, called for its side effect.
}
\description{
Show or hide a \code{\link[=bs_toast]{bs_toast()}} from server code.
}
\examples{
if (interactive()) show_bs_toast("hello")
if (interactive()) hide_bs_toast("hello")
}
