% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav-tabs.R
\name{update_bs_tabset}
\alias{update_bs_tabset}
\title{Control a tabset from the server}
\usage{
update_bs_tabset(id, selected, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Tabset id (namespaced automatically inside modules).}

\item{selected}{Value of the panel to show.}

\item{session}{The Shiny session.}
}
\value{
Invisibly \code{NULL}, called for its side effect.
}
\description{
Control a tabset from the server
}
\examples{
\dontrun{
update_bs_tabset("tabs", selected = "profile")
}
}
