% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{bs_close_button}
\alias{bs_close_button}
\title{Bootstrap close button}
\usage{
bs_close_button(..., label = "Close", white = FALSE, class = NULL)
}
\arguments{
\item{...}{Named HTML attributes (e.g. \code{data-bs-dismiss}).}

\item{label}{Accessible label.}

\item{white}{If \code{TRUE}, white variant for dark backgrounds.}

\item{class}{Extra classes.}
}
\value{
A button tag.
}
\description{
Bootstrap close button
}
