% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{bs_figure}
\alias{bs_figure}
\alias{bs_figure_img}
\alias{bs_figure_caption}
\title{Bootstrap figure}
\usage{
bs_figure(..., class = NULL)

bs_figure_img(src, ..., alt = NULL, class = NULL)

bs_figure_caption(..., class = NULL)
}
\arguments{
\item{...}{Figure content (image, caption) and named HTML attributes.}

\item{class}{Extra classes.}

\item{src}{Image source URL.}

\item{alt}{Alternative text.}
}
\value{
A figure tag.
}
\description{
Group an image with a caption. Compose \code{\link[=bs_figure]{bs_figure()}} with \code{\link[=bs_figure_img]{bs_figure_img()}}
and \code{\link[=bs_figure_caption]{bs_figure_caption()}}.
}
\examples{
bs_figure(
  bs_figure_img("photo.jpg", alt = "A photo"),
  bs_figure_caption("A caption for the image.")
)
}
