% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{bs_img}
\alias{bs_img}
\title{Bootstrap image}
\usage{
bs_img(
  src,
  ...,
  fluid = FALSE,
  thumbnail = FALSE,
  rounded = FALSE,
  alt = NULL,
  class = NULL
)
}
\arguments{
\item{src}{Image source URL.}

\item{...}{Extra named HTML attributes applied to the \verb{<img>}.}

\item{fluid}{Make the image responsive (\code{.img-fluid}).}

\item{thumbnail}{Render with a rounded thumbnail border (\code{.img-thumbnail}).}

\item{rounded}{Add rounded corners (\code{.rounded}).}

\item{alt}{Alternative text.}

\item{class}{Extra classes.}
}
\value{
An image tag.
}
\description{
Bootstrap image
}
\examples{
bs_img("logo.png", alt = "Logo", fluid = TRUE)
}
