% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{bs_password_input}
\alias{bs_password_input}
\title{Bootstrap password input}
\usage{
bs_password_input(
  id,
  label = NULL,
  value = "",
  ...,
  size = NULL,
  help = NULL,
  width = NULL
)
}
\arguments{
\item{id}{Input id; value available as \code{input$id}.}

\item{label}{Input label.}

\item{value}{Initial value.}

\item{...}{Extra attributes applied to the \verb{<input>} element.}

\item{size}{Control size: \code{"sm"} or \code{"lg"}.}

\item{help}{Help text rendered below the control (\code{.form-text}).}

\item{width}{CSS width (e.g. \code{"100\%"}, \code{"200px"}).}
}
\value{
A form control tag.
}
\description{
Bootstrap password input
}
\seealso{
\code{\link[shiny:passwordInput]{shiny::passwordInput()}}
}
